ajustes brutos na API

+++ src/main/java/br/gov/pf/model/dao/ExamDAO.java
@@ -5,7 +5,7 @@ import br.gov.pf.model.Exam;
 /**
  * Created by sartori on 13/07/17.
  */
-public class ExamDAO extends AbstractDAO<Integer, Exam> {
+public class ExamDAO extends AbstractDAO<Long, Exam> {
     public ExamDAO() {
         super();
     }
