+++ src/main/java/br/gov/pf/model/dao/GunDAO.java
@@ -5,7 +5,7 @@ import br.gov.pf.model.Gun;
 /**
  * Created by sartori on 13/07/17.
  */
-public class GunDAO extends AbstractDAO<Integer, Gun> {
+public class GunDAO extends AbstractDAO<Long, Gun> {
     public GunDAO() {
         super();
     }
