+++ src/main/java/br/gov/pf/model/dao/LicenceDAO.java
@@ -5,7 +5,7 @@ import br.gov.pf.model.Licence;
 /**
  * Created by sartori on 13/07/17.
  */
-public class LicenceDAO extends AbstractDAO<Integer, Licence> {
+public class LicenceDAO extends AbstractDAO<Long, Licence> {
     public LicenceDAO() {
         super();
     }
