+++ src/main/java/br/gov/pf/model/dao/UserDAO.java
@@ -5,7 +5,7 @@ import br.gov.pf.model.User;
 /**
  * Created by sartori on 13/07/17.
  */
-public class UserDAO extends AbstractDAO<Integer, User> {
+public class UserDAO extends AbstractDAO<Long, User> {
     public UserDAO() {
         super();
     }
