+++ src/main/java/br/gov/pf/model/service/ExamService.java
@@ -6,7 +6,7 @@ import br.gov.pf.model.dao.ExamDAO;
 /**
  * Created by sartori on 13/07/17.
  */
-public class ExamService extends AbstractService<Integer, Exam> {
+public class ExamService extends AbstractService<Long, Exam> {
     public ExamService() {
         dao = new ExamDAO();
     }
