+++ src/main/java/br/gov/pf/model/service/GunService.java
@@ -6,7 +6,7 @@ import br.gov.pf.model.dao.GunDAO;
 /**
  * Created by sartori on 13/07/17.
  */
-public class GunService extends AbstractService<Integer, Gun> {
+public class GunService extends AbstractService<Long, Gun> {
     public GunService() {
         dao = new GunDAO();
     }
