+++ src/main/java/br/gov/pf/model/service/LicenceService.java
@@ -6,7 +6,7 @@ import br.gov.pf.model.dao.LicenceDAO;
 /**
  * Created by sartori on 13/07/17.
  */
-public class LicenceService extends AbstractService<Integer, Licence>{
+public class LicenceService extends AbstractService<Long, Licence>{
     public LicenceService() {
         dao = new LicenceDAO();
     }
