+++ src/main/java/br/gov/pf/model/service/UserService.java
@@ -6,7 +6,7 @@ import br.gov.pf.model.dao.UserDAO;
 /**
  * Created by sartori on 13/07/17.
  */
-public class UserService extends AbstractService<Integer, User> {
+public class UserService extends AbstractService<Long, User> {
     public UserService() {
         dao = new UserDAO();
     }
